requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Django==4.1.3
django-bootstrap-v5==1.0.11
django-filter==22.1
django-jalali==6.0.1
django-jalali-date==1.0.1
django-webpush==0.3.4
djangorestframework==3.14.0
http-ece==1.1.0
idna==3.4
jdatetime==4.1.0
numpy==1.24.0
Pillow==9.3.0
psycopg2==2.9.5
py-vapid==1.9.0
pycparser==2.21
pytz==2022.6
pywebpush==1.9.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
xlwt==1.3.0
